Fixed the "Unwanted note on mute" bug
The reason why spacebar was toggling the buttons was because when you'd
click a button it would get focused, and cause spacebar to toggle it. So
now when the keyup function is called it checks to see if anything is
focused and it blurs it if it is.
I also made the Mute beat sound become red to match the mute toggle